requirements.txt
pandas==1.5.2
pandas-profiling==3.6.2
numpy==1.23.5
numpy-base==1.23.5
seaborn-base==0.12.2
streamlit==1.11.0
streamlit-pandas-profiling==0.1.3
matplotlib==3.6.0
matplotlib-base==3.6.0